requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
get-docker-secret==2.0.0
psycopg2==2.9.9
pytest==8.3.3
pytest-django==4.9.0